chore: move select test helpers to test-support new entrypoint in forms
