+++ packages/forms/package.json
@@ -51,6 +51,7 @@
     "ember-modifier": "^4.1.0"
   },
   "devDependencies": {
+    "@ember/test-helpers": "^4.0.4",
     "@babel/core": "7.26.0",
     "@babel/plugin-proposal-class-properties": "7.18.6",
     "@babel/plugin-proposal-decorators": "7.25.9",
